Add files via upload
If you do not know how to install Ionics in your Macintosh in the Terminal (or in the Command Prompt). This automate batch file will install the the "software" for youwithout typing any codes.

STEP 1:

Download the lastest Node.js from https://nodejs.org/en/

STEP 2: Unzip and open "ionic_installer.zip" file. This will automatically create a directory named "Ionic_Projects" in your Documents folder.

NOTE:

This is STILL A WORK IN PROGRESS. This is just the MAC INSTALLER FOR IONICS. I am will be soon working (when I'm not busy) on an automatic batch file where users can start, open, and create new ionic projects without typing codes in the Command Prompt or Terminal. But feel free to fork ahead!